Move embedder construction into a separate file
