+++ kglib/kgcn/pipeline/pipeline.py
@@ -22,8 +22,9 @@ import numpy as np
 from graph_nets.utils_np import graphs_tuple_to_networkxs
 
 from kglib.kgcn.learn.learn import KGCNLearner
-from kglib.kgcn.models.attribute import ContinuousAttribute, CategoricalAttribute, BlankAttribute
 from kglib.kgcn.models.core import softmax, KGCN
+from kglib.kgcn.pipeline.embed import construct_categorical_embedders, construct_continuous_embedders, \
+    construct_non_attribute_embedders
 from kglib.kgcn.pipeline.encode import encode_types, create_input_graph, create_target_graph, encode_values
 from kglib.kgcn.pipeline.utils import apply_logits_to_graphs, duplicate_edges_in_reverse
 from kglib.kgcn.plot.plotting import plot_across_training, plot_predictions
@@ -118,61 +119,3 @@ def pipeline(graphs,
 
     _, _, _, _, _, solveds_tr, solveds_ge = tr_info
     return ge_graphs, solveds_tr, solveds_ge
-
-
-def construct_categorical_embedders(node_types, attr_embedding_dim, categorical_attributes):
-    attr_embedders = dict()
-
-    # Construct attribute embedders
-    for attribute_type, categories in categorical_attributes.items():
-
-        attr_typ_index = node_types.index(attribute_type)
-
-        def make_embedder():
-            return CategoricalAttribute(len(categories), attr_embedding_dim,
-                                        name=attribute_type + '_cat_embedder')
-
-        # Record the embedder, and the index of the type that it should encode
-        attr_embedders[make_embedder] = [attr_typ_index]
-
-    return attr_embedders
-
-
-def construct_continuous_embedders(node_types, attr_embedding_dim, continuous_attributes):
-    attr_embedders = dict()
-
-    # Construct attribute embedders
-    for attribute_type in continuous_attributes.keys():
-
-        attr_typ_index = node_types.index(attribute_type)
-
-        def make_embedder():
-            return ContinuousAttribute(attr_embedding_dim, name=attribute_type + '_cat_embedder')
-
-        # Record the embedder, and the index of the type that it should encode
-        attr_embedders[make_embedder] = [attr_typ_index]
-
-    return attr_embedders
-
-
-def construct_non_attribute_embedders(node_types, attr_embedding_dim, categorical_attributes, continuous_attributes):
-
-    attribute_names = list(categorical_attributes.keys())
-    attribute_names.extend(list(continuous_attributes.keys()))
-
-    non_attribute_nodes = []
-    for i, type in enumerate(node_types):
-        if type not in attribute_names:
-            non_attribute_nodes.append(i)
-
-    # All entities and relations (non-attributes) also need an embedder with matching output dimension, which does
-    # nothing. This is provided as a list of their indices
-    def make_blank_embedder():
-        return BlankAttribute(attr_embedding_dim)
-
-    attr_embedders = dict()
-
-    if len(non_attribute_nodes) > 0:
-        attr_embedders[make_blank_embedder] = non_attribute_nodes
-    return attr_embedders
-
